+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- feat(databases): add search parameter to list endpoint
 - chore(databases): make pagination fields nullable
 
 ### Removed